RefAllocator fix: referencing the same allocator was wrong
Handling of referencing the same allocator twice was not implemented correctly.
The list of currently tracked allocators is now searched for the matching one
and its refcount is incremented accordingly.

Tests were adapted as they contained misleading or even wrong code in dealing
with allocators.

Thread was corrected in some places to not allocate any long living objects on
fAllocator as this could lead to incorrect memory accesses when using an
allocator during Thread.Start which differs from the one used to construct the
thread.

Change-Id: I7bfcf9607dc33592690581cdb2c729f4077aa0e1